% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oas_countries.R
\name{get_oas_countries}
\alias{get_oas_countries}
\title{Get OAS countries}
\usage{
get_oas_countries()
}
\value{
A dataset with one column iso3c
}
\description{
Get OAS countries
}
\examples{
get_oas_countries() \%>\% dplyr::pull(iso3c)
library(tidyverse)
get_oas_countries() \%>\% dplyr::pull(iso3c)
}
